Justarade basen litegrann.
