+++ lib/map_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class MapPage extends StatelessWidget {
+  const MapPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("This is the map page."),
+      ),
+    );
+  }
+}
