+++ lib/social_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class SettingsPage extends StatelessWidget {
+  const SettingsPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("This is the settings page."),
+      ),
+    );
+  }
+}
