+++ lib/sounds_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class SharePage extends StatelessWidget {
+  const SharePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("This is the share page."),
+      ),
+    );
+  }
+}
